convnext/requirements.txt
{\rtf1\ansi\ansicpg1252\cocoartf2639
\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fmodern\fcharset0 Courier-Oblique;\f1\fmodern\fcharset0 Courier;}
{\colortbl;\red255\green255\blue255;\red81\green136\blue67;\red32\green32\blue32;\red153\green168\blue186;
}
{\*\expandedcolortbl;;\csgenericrgb\c31765\c53333\c26275;\csgenericrgb\c12549\c12549\c12549;\csgenericrgb\c60000\c65882\c72941;
}
\margl1440\margr1440\vieww11520\viewh8400\viewkind0
\pard\tx560\tx1120\tx1680\tx2240\tx2800\tx3360\tx3920\tx4480\tx5040\tx5600\tx6160\tx6720\pardirnatural\partightenfactor0

\f0\i\fs26 \cf2 \cb3 # This file may be used to create an environment using:\
# $ conda create --name <env> --file <this file>\
# platform: linux-64\

\f1\i0 \cf4 _libgcc_mutex=0.1=main\
_openmp_mutex=5.1=1_gnu\
absl-py=1.4.0=pypi_0\
addict=2.4.0=pypi_0\
blessed=1.20.0=pypi_0\
ca-certificates=2023.01.10=h06a4308_0\
cachetools=5.3.0=pypi_0\
certifi=2022.12.7=py38h06a4308_0\
charset-normalizer=3.1.0=pypi_0\
cloudpickle=2.2.1=pypi_0\
cmake=3.26.3=pypi_0\
contourpy=1.0.7=pypi_0\
cycler=0.11.0=pypi_0\
einops=0.6.0=pypi_0\
filelock=3.12.0=pypi_0\
fonttools=4.39.2=pypi_0\
google-auth=2.17.3=pypi_0\
google-auth-oauthlib=1.0.0=pypi_0\
gpustat=1.1=pypi_0\
grpcio=1.53.0=pypi_0\
huggingface-hub=0.13.4=pypi_0\
idna=3.4=pypi_0\
importlib-metadata=6.4.1=pypi_0\
importlib-resources=5.12.0=pypi_0\
jinja2=3.1.2=pypi_0\
kiwisolver=1.4.4=pypi_0\
ld_impl_linux-64=2.38=h1181459_1\
libffi=3.4.2=h6a678d5_6\
libgcc-ng=11.2.0=h1234567_1\
libgomp=11.2.0=h1234567_1\
libstdcxx-ng=11.2.0=h1234567_1\
lit=16.0.1=pypi_0\
markdown=3.4.3=pypi_0\
markupsafe=2.1.2=pypi_0\
matplotlib=3.7.1=pypi_0\
mmcls=0.25.0=pypi_0\
mmcv-full=1.7.1=pypi_0\
mmsegmentation=0.30.0=pypi_0\
mpmath=1.3.0=pypi_0\
ncurses=6.4=h6a678d5_0\
networkx=3.1=pypi_0\
numpy=1.24.2=pypi_0\
nvidia-cublas-cu11=11.10.3.66=pypi_0\
nvidia-cuda-cupti-cu11=11.7.101=pypi_0\
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0\
nvidia-cuda-runtime-cu11=11.7.99=pypi_0\
nvidia-cudnn-cu11=8.5.0.96=pypi_0\
nvidia-cufft-cu11=10.9.0.58=pypi_0\
nvidia-curand-cu11=10.2.10.91=pypi_0\
nvidia-cusolver-cu11=11.4.0.1=pypi_0\
nvidia-cusparse-cu11=11.7.4.91=pypi_0\
nvidia-ml-py=11.525.112=pypi_0\
nvidia-nccl-cu11=2.14.3=pypi_0\
nvidia-nvtx-cu11=11.7.91=pypi_0\
oauthlib=3.2.2=pypi_0\
opencv-python=4.7.0.72=pypi_0\
openssl=1.1.1t=h7f8727e_0\
packaging=23.0=pypi_0\
pillow=9.4.0=pypi_0\
pip=23.0.1=py38h06a4308_0\
prettytable=3.6.0=pypi_0\
protobuf=3.20.3=pypi_0\
psutil=5.9.5=pypi_0\
pyasn1=0.4.8=pypi_0\
pyasn1-modules=0.2.8=pypi_0\
pyparsing=3.0.9=pypi_0\
python=3.8.16=h7a1cb2a_3\
python-dateutil=2.8.2=pypi_0\
pyyaml=6.0=pypi_0\
readline=8.2=h5eee18b_0\
requests=2.28.2=pypi_0\
requests-oauthlib=1.3.1=pypi_0\
rsa=4.9=pypi_0\
scipy=1.10.1=pypi_0\
setuptools=65.6.3=py38h06a4308_0\
six=1.16.0=pypi_0\
sqlite=3.41.1=h5eee18b_0\
submitit=1.4.5=pypi_0\
sympy=1.11.1=pypi_0\
tabulate=0.9.0=pypi_0\
tensorboard=2.12.2=pypi_0\
tensorboard-data-server=0.7.0=pypi_0\
tensorboard-plugin-wit=1.8.1=pypi_0\
tensorboardx=2.6=pypi_0\
timm=0.6.13=pypi_0\
tk=8.6.12=h1ccaba5_0\
torch=1.12.0+cu116=pypi_0\
torchaudio=0.12.0+cu116=pypi_0\
torchvision=0.13.0+cu116=pypi_0\
tqdm=4.65.0=pypi_0\
triton=2.0.0=pypi_0\
typing-extensions=4.5.0=pypi_0\
urllib3=1.26.15=pypi_0\
wcwidth=0.2.6=pypi_0\
werkzeug=2.2.3=pypi_0\
wheel=0.38.4=py38h06a4308_0\
xz=5.2.10=h5eee18b_1\
yapf=0.32.0=pypi_0\
zipp=3.15.0=pypi_0\
zlib=1.2.13=h5eee18b_0\
\
}